eegplot window refreshes were slow
Window refreshes were slow, sometimes taking 47 seconds for my data.
Changed calls to axes() which involve redrawing the window, to setting the current axes of the figure, which is a passive change and does not redraw the window.
Window renders now take about 9 seconds, most of which is spent in plot().