Update pop_saveset.m
Sorry if I made this suggestion already but I don't see that I did anywhere. This line gives unexpected behavior of not saving a set when it deems it unnecessary. The thing is it doesn't detect any changes made via command line or anything that doesn't also update the EEG.saved property. I think the risk of saving unnecessarily is preferable to the risk of losing information from not saving unexpectedly.