Add Octave support to plugin installation
Due to java calls in `plugin_urlread` and `plugin_urlwrite` functions
and the use of libraries `com.mathworks.mlwidgets.io.InterruptibleStreamCopier`
and `com.mathworks.mlwidgets.html.HTMLPrefs.setProxySettings`, `plugin_install`
and `plugin_askinstall` cannot be used with GNU Octave.
This is solved by using the built-in `urlread` and `urlwrite` functions
if Octave is detected. The textscan format in `plugin_getweb` is also
updated to correctly parse the plugin sizes if they are succeeded by an
optional 'KB' or 'MB', which is handled differently in Octave and in
Matlab. Currently, these 'KB' and 'MB' size specifications are ignored,
which was also the case earlier.